+++ filament/src/VertexBuffer.cpp
@@ -221,7 +221,7 @@ void VertexBuffer::setBufferAt(Engine& engine, uint8_t bufferIndex,
 }
 
 void VertexBuffer::populateTangentQuaternions(const QuatTangentContext& ctx) {
-    auto quats = geometry::SurfaceOrientation::Builder()
+    auto* quats = geometry::SurfaceOrientation::Builder()
         .vertexCount(ctx.quatCount)
         .normals(ctx.normals, ctx.normalsStride)
         .tangents(ctx.tangents, ctx.tangentsStride)
@@ -229,15 +229,17 @@ void VertexBuffer::populateTangentQuaternions(const QuatTangentContext& ctx) {
 
     switch (ctx.quatType) {
         case HALF4:
-            quats.getQuats((quath*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
+            quats->getQuats((quath*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
             break;
         case SHORT4:
-            quats.getQuats((short4*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
+            quats->getQuats((short4*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
             break;
         case FLOAT4:
-            quats.getQuats((quatf*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
+            quats->getQuats((quatf*) ctx.outBuffer, ctx.quatCount, ctx.outStride);
             break;
     }
+
+    delete quats;
 }
 
 } // namespace filament