+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ A new header is inserted each time a *tag* is created.
 - gltfio: generate normals for flat-shaded models that do not have normals.
 - Material instances now allow dynamic depth testing and other rasterization state.
 - Support for Bloom as a post-process effect.
+- Added Java bindings for geometry::SurfaceOrientation.
 
 ## v1.4.5